+++ apps/member-profile/app/routes/_profile.events.tsx
@@ -1,9 +1,10 @@
 import { json, type LoaderFunctionArgs } from '@remix-run/node';
-import { NavLink, Outlet, useLoaderData } from '@remix-run/react';
+import { Outlet, useLoaderData } from '@remix-run/react';
 
-import { cx, Text } from '@oyster/ui';
+import { Text } from '@oyster/ui';
 
 import { countPastEvents, countUpcomingEvents } from '@/member-profile.server';
+import { NavigationItem } from '@/shared/components/navigation';
 import { Route } from '@/shared/constants';
 import { ensureUserAuthenticated } from '@/shared/session.server';
 
@@ -30,39 +31,15 @@ export default function EventsLayout() {
       <Text variant="2xl">Events 📅</Text>
 
       <ul className="flex gap-4">
-        <EventNavigationItem
-          label={`Upcoming (${upcomingEventsCount})`}
-          to={Route['/events/upcoming']}
-        />
-        <EventNavigationItem
-          label={`Past (${pastEventsCount})`}
-          to={Route['/events/past']}
-        />
+        <NavigationItem to={Route['/events/upcoming']}>
+          Upcoming ({upcomingEventsCount})
+        </NavigationItem>
+
+        <NavigationItem to={Route['/events/past']}>
+          Past ({pastEventsCount})
+        </NavigationItem>
       </ul>
       <Outlet />
     </>
-  );
-}
-
-type EventNavigationItemProps = {
-  label: string;
-  to: string;
-};
-
-function EventNavigationItem({ label, to }: EventNavigationItemProps) {
-  return (
-    <li>
-      <NavLink
-        className={({ isActive }) => {
-          return cx(
-            'underline hover:text-primary',
-            isActive && 'text-primary underline'
-          );
-        }}
-        to={to}
-      >
-        {label}
-      </NavLink>
-    </li>
   );
 }
