+++ apps/member-profile/app/routes/_profile.weekly-recap.$date._index.tsx
@@ -1,0 +1,12 @@
+import { type LoaderFunctionArgs, redirect } from '@remix-run/node';
+import { generatePath } from '@remix-run/react';
+
+import { Route } from '@/shared/constants';
+
+export async function loader({ params }: LoaderFunctionArgs) {
+  return redirect(
+    generatePath(Route['/weekly-recap/:date/leaderboard'], {
+      date: params.date as string,
+    })
+  );
+}
