+++ apps/member-profile/app/routes/_profile.weekly-recap.$date.tsx
@@ -1,26 +1,34 @@
 import { json, type LoaderFunctionArgs } from '@remix-run/node';
-import { useLoaderData } from '@remix-run/react';
+import {
+  generatePath,
+  Outlet,
+  useLoaderData,
+  useParams,
+} from '@remix-run/react';
 import dayjs from 'dayjs';
+import { type PropsWithChildren } from 'react';
 
 import { listCompanyReviews } from '@oyster/core/employment.server';
 import { listResources } from '@oyster/core/resources.server';
 import { listSlackMessages } from '@oyster/core/slack.server';
-import { getPresignedURL } from '@oyster/infrastructure/object-storage';
 import { Divider, Text } from '@oyster/ui';
+import { iife } from '@oyster/utils';
 
-import { getPointsLeaderboard } from '@/member-profile.server';
-import { type EmploymentType, type LocationType } from '@/member-profile.ui';
-import { type ResourceType } from '@/modules/resource';
-import { Card } from '@/shared/components/card';
-import { CompanyReview } from '@/shared/components/company-review';
-import { Leaderboard } from '@/shared/components/leaderboard';
-import { Resource } from '@/shared/components/resource';
-import { SlackMessage } from '@/shared/components/slack-message';
-import { getTimezone } from '@/shared/cookies.server';
-import { ensureUserAuthenticated, user } from '@/shared/session.server';
+import { NavigationItem } from '@/shared/components/navigation';
+import { type Route } from '@/shared/constants';
+import { ensureUserAuthenticated } from '@/shared/session.server';
+
+export function getDateRange(date: unknown) {
+  const dayObject = dayjs(date as string);
+
+  return {
+    startOfWeek: dayObject.startOf('week').toDate(),
+    endOfWeek: dayObject.endOf('week').toDate(),
+  };
+}
 
 export async function loader({ params, request }: LoaderFunctionArgs) {
-  const session = await ensureUserAuthenticated(request);
+  await ensureUserAuthenticated(request);
 
   const dayObject = dayjs(params.date);
 
@@ -28,280 +36,62 @@ export async function loader({ params, request }: LoaderFunctionArgs) {
     throw new Response('Not a valid date.', { status: 400 });
   }
 
-  const startOfWeek = dayObject.startOf('week').toDate();
-  const endOfWeek = dayObject.endOf('week').toDate();
+  const { endOfWeek, startOfWeek } = getDateRange(params.date);
 
-  const memberId = user(session);
+  const [announcements, reviews, { totalCount: totalResources }] =
+    await Promise.all([
+      listSlackMessages({
+        pagination: { page: 1, limit: 10 },
+        select: ['messages.id'],
+        where: {
+          channelId: '', // Announcements Channel ID
+          sentAfter: startOfWeek,
+          sentBefore: endOfWeek,
+        },
+      }),
 
-  const [
-    announcementMessages,
-    topMessages,
-    leaderboard,
-    _reviews,
-    { resources: _resources },
-  ] = await Promise.all([
-    listSlackMessages({
-      include: ['reactions'],
-      pagination: {
-        page: 1,
-        limit: 10,
-      },
-      select: ['messages.id', 'messages.text'],
-      where: {
-        channelId: '', // Announcements Channel ID
-        sentAfter: startOfWeek,
-        sentBefore: endOfWeek,
-      },
-    }),
+      listCompanyReviews({
+        select: ['companyReviews.id'],
+        where: {
+          postedAfter: startOfWeek,
+          postedBefore: endOfWeek,
+        },
+      }),
 
-    listSlackMessages({
-      include: ['reactions'],
-      select: ['messages.id', 'messages.text'],
-      orderBy: 'most_reactions',
-      pagination: {
-        page: 1,
-        limit: 10,
-      },
-      where: {
-        sentAfter: startOfWeek,
-        sentBefore: endOfWeek,
-      },
-    }),
+      listResources({
+        memberId: '',
+        pagination: { limit: 1, page: 1 },
+        orderBy: 'newest',
+        select: [],
+        where: {
+          postedAfter: startOfWeek,
+          postedBefore: endOfWeek,
+          search: '',
+          tags: [],
+        },
+      }),
+    ]);
 
-    getPointsLeaderboard({
-      limit: 25,
-      where: {
-        memberId,
-        occurredAfter: startOfWeek,
-        occurredBefore: endOfWeek,
-      },
-    }),
+  const dateRange = iife(() => {
+    const format = 'dddd, MMMM D, YYYY';
 
-    listCompanyReviews({
-      select: [
-        'companyReviews.createdAt',
-        'companyReviews.id',
-        'companyReviews.rating',
-        'companyReviews.recommend',
-        'companyReviews.text',
-        'students.id as reviewerId',
-        'students.firstName as reviewerFirstName',
-        'students.lastName as reviewerLastName',
-        'students.profilePicture as reviewerProfilePicture',
-        'workExperiences.employmentType',
-        'workExperiences.endDate',
-        'workExperiences.locationCity',
-        'workExperiences.locationState',
-        'workExperiences.locationType',
-        'workExperiences.startDate',
-        'workExperiences.title',
-      ],
-      where: {
-        postedAfter: startOfWeek,
-        postedBefore: endOfWeek,
-      },
-    }),
+    const startRange = dayObject.startOf('week').format(format);
+    const endRange = dayObject.endOf('week').format(format);
 
-    listResources({
-      memberId: user(session),
-      orderBy: 'most_upvotes',
-      pagination: {
-        limit: 1000,
-        page: 1,
-      },
-      select: [
-        'resources.description',
-        'resources.id',
-        'resources.link',
-        'resources.postedAt',
-        'resources.title',
-        'resources.type',
-        'students.firstName as posterFirstName',
-        'students.id as posterId',
-        'students.lastName as posterLastName',
-        'students.profilePicture as posterProfilePicture',
-      ],
-      where: {
-        postedAfter: startOfWeek,
-        postedBefore: endOfWeek,
-        search: '',
-        tags: [],
-      },
-    }),
-  ]);
-
-  const reviews = _reviews.map(
-    ({ createdAt, endDate, startDate, ...review }) => {
-      const startMonth = dayjs.utc(startDate).format('MMMM YYYY');
-
-      const endMonth = endDate
-        ? dayjs.utc(endDate).format('MMMM YYYY')
-        : 'Present';
-
-      return {
-        ...review,
-        date: `${startMonth} - ${endMonth}`,
-        reviewedAt: dayjs().to(createdAt),
-      };
-    }
-  );
-
-  const url = new URL(request.url);
-
-  const resources = await Promise.all(
-    _resources.map(
-      async ({
-        attachments,
-        postedAt,
-        tags,
-        upvotes,
-        upvoted,
-        views,
-        ...record
-      }) => {
-        return {
-          ...record,
-
-          // If there are any attachments, we need to generate a presigned URL
-          // to the object in the Cloudflare R2 bucket.
-          attachments: await Promise.all(
-            (attachments || []).map(async (attachment) => {
-              return {
-                mimeType: attachment.mimeType,
-                uri: await getPresignedURL({
-                  expiresIn: 60 * 60, // 1 hour
-                  key: attachment.s3Key,
-                }),
-              };
-            })
-          ),
-
-          // If the logged-in member is the poster of the resource, they should
-          // be able to edit the resource.
-          editable: record.posterId === user(session),
-
-          // This is a relative time of when the resource was posted,
-          // ie: "2d" (2 days ago).
-          postedAt: dayjs().to(postedAt),
-
-          postedAtExpanded: dayjs(postedAt)
-            .tz(getTimezone(request))
-            .format('MMM DD, YYYY • h:mm A'),
-
-          // This is the URL that can be shared with others to view the
-          // resource. Note: This is a URL to our application, not the _actual_
-          // resource URL (which has permissions via the presigned URL).
-          shareableUri: `${url.protocol}//${url.host}/resources?id=${record.id}`,
-
-          tags: tags!,
-          upvotes: Number(upvotes),
-          upvoted: Boolean(upvoted),
-          views: Number(views),
-        };
-      }
-    )
-  );
-
-  const fakeAnnouncementMessages: {
-    id: string;
-    text: string | null;
-    // totalReactions?: string | null | undefined;
-  }[] = [
-    {
-      id: '1',
-      text: `<!channel> Hey fam.
-
-As promised, I wanted to address yesterday’s events regarding ColorStack and the termination of a now former employee.
-
-Yesterday, we had to make the difficult decision to part ways with one of our team members. I believe in handling these matters confidentially, but unfortunately, they chose to share details of a private conversation, including an audio recording, on LinkedIn.
-
-I want to reassure you that while this situation is unfortunate, it does not reflect any change in our team’s focus. Every decision that I’ve made up until this point has been with you all in mind, to build the team that will help us fulfill our mission effectively. We remain fully dedicated to this.
-
-We are addressing this matter internally, but rest assured there will be no disruptions to any of our programs and upcoming events. We have a coverage plan in place to continue to attract and retain corporate partners that want to hire you all until we backfill the role.
-
-Building a company is hard. There is always a lesson to be learned or an area for growth. This will only make us better as an organization moving forward.
-
-Thank you for your understanding and continued trust in ColorStack.
-
--Jehron`,
-    },
-    {
-      id: '2',
-      text: ` <!channel> Hey Y'all! Happy Friday!! Keeping this one short and sweet!
-
- *Father's Day Celebration!* :green_heart: :fist::skin-tone-4:
- :sparkles: With Dad's Day this Sunday, we wondered if we have any Dads in our ColorStack Family?!
- We want to celebrate you!
- :arrow_right: *<https://docs.google.com/forms/d/e/1FAIpQLSc4kcMfg2YWhkC701iJQpEWBzeVHiTkfGNqJTEkhufhEnKncw/viewform|Fill out this short form >*to share how ColorStack has impacted your journey as a student/young professional and a father!
-
- *Member Profile New Releases!* :colorstack_logo:
- :sparkles: If you are wondering if other ColorStack'ers work at or have had a good experience with a company you're interested in, check out our:
- :arrow_right: <https://app.colorstack.io/companies|Companies Feature>
-
- :sparkles: Don't forget about our <https://app.colorstack.io/resources|Resource Database>! There is so much useful information, tools, and resources to guide and support you on your journeys in ColorStack, at your workplaces, during your job search, or in your classroom endeavors!
- :arrow_right: <https://app.colorstack.io/resources|Resource Database>
-
- *QOTD: Where do you dream of living after graduating college?*- <@U0730BL4HQE>
- Itching to ask the ColorStack Community something? <https://forms.gle/mVFZoWo2XW8z39HYA|Submit a QOTD here>`,
-    },
-    {
-      id: '3',
-      text: `<!channel> Happy Fridayyy!! Happy first end of week to those who started your internships or full-time roles this week! :party_blob: :party_blob:
-This is a long one, but lots of *GOOD* info below!
-
-*Alumni Channels Launch!*
-:sparkles: Y'all have asked and asked....they are finally HERE! Alumni channels have launched and are ready for you to join! These channels are v1 of a MUCH larger alumni strategy still in the works. However, to give our alumni members places to chat, connect, and support one another in transitions- we are releasing channels today!
-
-<#C06PXL47X5L|alumni-announcements> <#C07750LMRSN|alumni-coding-help>, <#C0774SBJPQB|alumni-finance>, <#C07750JGHM0|alumni-housing>, <#C076YBA33SA|alumni-office-culture>, <#C06QGUH5GP3|alumni-questions>, <#C06QGUE6TJM|alumni-random> are ready for you to join and share!
-
-Currently, these channels are public and not restricted to only alumni. However, let's respect their purpose and give our alumni these spaces for connection and growth.
-
-*Why Alumni Channels?- Doesn't this divide the community?*
-:heavy_check_mark: Yes, it does. But this division is strategic. As ColorStack members transition into full-time opportunities, staying active in the community has become challenging. We have former members doing incredible things in the industry and their lives, and we're missing out.
-:handshake::skin-tone-4: _The alumni channels are designed to bridge this gap and keep us connected._
-
-:sparkles: We want to create a world WITHIN the ColorStack community, where our alums can be active, engaged, and plugged into one another without the noise and disruption of our larger Slack community.
-
-Remember, this is just the beginning. Launching these channels is the first phase of a much larger alumni strategy. In the future, we plan to provide opportunities for alumni to give back to the community and be valuable resources.
-:party_blob: _Stay tuned; exciting things are on the horizon!_
-
-*Opportunities!!* :moneybag: :moneybag:
-Several opportunities & events from our partners were shared in our channels this week! SO many ColorStack members have gotten their roles or flown to all-expense paid events because of opportunities they've learned about in this community just like these. Don't miss your chance!!
-
-• Apply, and be sure to let them know ColorStack sent you!! :saluting_face:
-1. <https://colorstack-family.slack.com/archives/C011H0EFU14/p1717689602203669|Bill 2025 Opportunities>
-2. <https://colorstack-family.slack.com/archives/C011H0EFU14/p1717624800382459|Salesforce Tech Equality Summit>  :rotating_light: Deadline *TODAY*
-3. <https://colorstack-family.slack.com/archives/C011H0EFU14/p1717614000393319|Jane Street>  :rotating_light: Deadline *TODAY*
-*QOTD: What is the best 2 snack combo out there? <@U074DVD3QUR>*
-Itching to ask the ColorStack Community something? <https://forms.gle/mVFZoWo2XW8z39HYA|Submit a QOTD here>`,
-    },
-  ];
+    return `${startRange} - ${endRange}`;
+  });
 
   return json({
-    announcementMessages: fakeAnnouncementMessages,
-    dateRange: getDateRange(params.date as string),
-    leaderboard,
-    resources,
-    reviews,
-    topMessages,
+    dateRange,
+    totalAnnouncements: announcements.length,
+    totalResources,
+    totalReviews: reviews.length,
   });
 }
 
-function getDateRange(date: string) {
-  const dayObject = dayjs(date);
-
-  const format = 'dddd, MMMM D, YYYY';
-
-  const dateRange =
-    dayObject.startOf('week').format(format) +
-    ' - ' +
-    dayObject.endOf('week').format(format);
-
-  return dateRange;
-}
-
-export default function WeekInReviewPage() {
-  const { dateRange } = useLoaderData<typeof loader>();
+export default function WeekInReviewLayout() {
+  const { dateRange, totalResources, totalReviews, totalAnnouncements } =
+    useLoaderData<typeof loader>();
 
   return (
     <section className="mx-auto flex w-full max-w-[36rem] flex-col gap-4 @container">
@@ -310,172 +100,58 @@ export default function WeekInReviewPage() {
         <Text color="gray-500">{dateRange}</Text>
       </header>
 
-      <PointsLeaderboard />
-      <AnnouncementMessages />
-      <TopMessages />
-      <Resources />
-      <CompanyReviews />
+      <nav>
+        <ul className="flex flex-wrap gap-x-4 gap-y-2">
+          <RecapNavigationItem to="/weekly-recap/:date/leaderboard">
+            Leaderboard
+          </RecapNavigationItem>
+          <RecapNavigationItem to="/weekly-recap/:date/announcements">
+            Announcements ({totalAnnouncements})
+          </RecapNavigationItem>
+          <RecapNavigationItem to="/weekly-recap/:date/resources">
+            Resources ({totalResources})
+          </RecapNavigationItem>
+          <RecapNavigationItem to="/weekly-recap/:date/reviews">
+            Company Reviews ({totalReviews})
+          </RecapNavigationItem>
+        </ul>
+      </nav>
+
+      <Divider my="2" />
+      <Outlet />
     </section>
   );
 }
 
-function AnnouncementMessages() {
-  const { announcementMessages } = useLoaderData<typeof loader>();
+function RecapNavigationItem({
+  children,
+  to,
+}: PropsWithChildren<{ to: Route }>) {
+  const { date } = useParams();
 
   return (
-    <Card>
-      <Card.Title>Announcements ({announcementMessages.length})</Card.Title>
-
-      <Card.Description>
-        Announcements from the ColorStack team this week in #announcements.
-      </Card.Description>
-
-      <ul className="flex flex-col gap-4">
-        {announcementMessages.map((message) => {
-          return (
-            <>
-              <Divider />
-              <SlackMessage key={message.id}>{message.text}</SlackMessage>
-            </>
-          );
-        })}
-      </ul>
-    </Card>
+    <NavigationItem to={generatePath(to as string, { date })}>
+      {children}
+    </NavigationItem>
   );
 }
 
-function PointsLeaderboard() {
-  const { leaderboard } = useLoaderData<typeof loader>();
+// Shared
 
+type RecapSectionProps = PropsWithChildren<{
+  description: React.ReactNode;
+  title: React.ReactNode;
+}>;
+
+export function RecapPage({ children, description, title }: RecapSectionProps) {
   return (
-    <Card>
-      <Card.Title>Points Leaderboard 🏆</Card.Title>
+    <section className="flex flex-col gap-[inherit]">
+      <header className="flex flex-col gap-1">
+        <Text variant="xl">{title}</Text>
+        <Text color="gray-500">{description}</Text>
+      </header>
 
-      <Card.Description>
-        The top point earners in the ColorStack Family this week.
-      </Card.Description>
-
-      <Leaderboard.List>
-        {leaderboard.map((position) => {
-          return (
-            <Leaderboard.Item
-              key={position.id}
-              firstName={position.firstName}
-              label={<LeaderboardItemLabel points={position.points} />}
-              lastName={position.lastName}
-              me={position.me}
-              position={position.rank}
-              profilePicture={position.profilePicture || undefined}
-            />
-          );
-        })}
-      </Leaderboard.List>
-    </Card>
+      {children}
+    </section>
   );
-}
-
-function LeaderboardItemLabel({ points }: { points: number }) {
-  const formatter = Intl.NumberFormat('en-US');
-
-  return (
-    <Leaderboard.ItemLabel>
-      {formatter.format(points)} <span className="text-sm"> Points</span>
-    </Leaderboard.ItemLabel>
-  );
-}
-
-function Resources() {
-  const { resources } = useLoaderData<typeof loader>();
-
-  return (
-    <Card>
-      <Card.Title>Resources 📚 ({resources.length})</Card.Title>
-
-      <Card.Description>
-        Helpful resources that were shared this week.
-      </Card.Description>
-
-      <Resource.List>
-        {resources.map((resource) => {
-          return (
-            <>
-              <Divider />
-
-              <Resource
-                key={resource.id}
-                attachments={resource.attachments}
-                border={false}
-                description={resource.description}
-                editable={resource.editable}
-                id={resource.id}
-                link={resource.link}
-                postedAt={resource.postedAt}
-                postedAtExpanded={resource.postedAtExpanded}
-                posterFirstName={resource.posterFirstName as string}
-                posterId={resource.posterId as string}
-                posterLastName={resource.posterLastName as string}
-                posterProfilePicture={resource.posterProfilePicture}
-                shareableUri={resource.shareableUri}
-                tags={resource.tags}
-                title={resource.title}
-                type={resource.type as ResourceType}
-                upvoted={resource.upvoted}
-                upvotes={resource.upvotes}
-                views={resource.views}
-              />
-            </>
-          );
-        })}
-      </Resource.List>
-    </Card>
-  );
-}
-
-function CompanyReviews() {
-  const { reviews } = useLoaderData<typeof loader>();
-
-  return (
-    <Card>
-      <Card.Title>Company Reviews ({reviews.length})</Card.Title>
-
-      <Card.Description>
-        See what your peers have to say about their recent work experiences!
-      </Card.Description>
-
-      <CompanyReview.List>
-        {reviews.map((review) => {
-          return (
-            <>
-              <Divider />
-
-              <CompanyReview
-                key={review.id}
-                border={false}
-                date={review.date}
-                employmentType={review.employmentType as EmploymentType}
-                locationCity={review.locationCity}
-                locationState={review.locationState}
-                locationType={review.locationType as LocationType}
-                rating={review.rating}
-                recommend={review.recommend}
-                reviewedAt={review.reviewedAt}
-                reviewerFirstName={review.reviewerFirstName as string}
-                reviewerId={review.reviewerId as string}
-                reviewerLastName={review.reviewerLastName as string}
-                reviewerProfilePicture={review.reviewerProfilePicture}
-                text={review.text}
-                title={review.title as string}
-              />
-            </>
-          );
-        })}
-      </CompanyReview.List>
-    </Card>
-  );
-}
-
-function TopMessages() {
-  const { topMessages } = useLoaderData<typeof loader>();
-
-  return null;
 }
