+++ apps/member-profile/app/shared/constants.ts
@@ -48,6 +48,10 @@ const ROUTES = [
   '/resources',
   '/resources/add',
   '/resources/:id/edit',
+  '/weekly-recap/:date/announcements',
+  '/weekly-recap/:date/leaderboard',
+  '/weekly-recap/:date/resources',
+  '/weekly-recap/:date/reviews',
 ] as const;
 
 export type Route = (typeof ROUTES)[number];
